split in modules acceleration, angular acceleration, angular velocity
